All seems to work
